Change standalone flag of binstubs cmd to boolean
This flag was introduced in #4610 declared with the same code as the flag
found in the `install` command.

In the case of the `install` command, it makes sense to be an Array option, since
one can specify groups.

This doesn't hold true for the `binstubs` cmd where `--standalone` stands to signify
whether the generated stubs will require the Bundler runtime or not.